change the arg of move method, using env for user and pw

+++ rms/addproductdetails.py
@@ -106,5 +106,5 @@ class AddProductDetails(QFrame):
         '''centers the window on the screen'''
         screen = QDesktopWidget().screenGeometry()
         size = self.geometry()
-        self.move((screen.width() - size.width()) / 2,
-                  (screen.height() - size.height()) / 2)
+        self.move(int((screen.width() - size.width()) / 2),
+                  int((screen.height() - size.height()) / 2))
